make star index unsigned int

+++ src/parse_stars/star.rs
@@ -2,7 +2,7 @@
 #[derive(Debug)]
 pub struct Star {
     /// the index of the star, serving as its name. Stars are to be refered to by index
-    pub index: i32,
+    pub index: u16,
     /// the x, y, z coordinates of a star in the earth reference frame
     pub coords: [f64; 3], 
     /// the visual brigthness of the star in magnitudes
@@ -10,7 +10,7 @@ pub struct Star {
 }
 
 impl Star {
-    pub fn new(ra: f64, dec: f64, brightness: f64, index: i32) -> Star {
+    pub fn new(ra: f64, dec: f64, brightness: f64, index: u16) -> Star {
         Star {
             index,
             coords: [dec.cos()*ra.cos(), dec.cos()*ra.sin(), dec.sin()],
